update settings and help menu
