+++ src/types/settings.ts
@@ -1,10 +1,6 @@
 export interface ISettingsProps {
   actions?: {
-    signOut: Function;
     changeTheme: Function;
     consentToScopes: Function;
-  };
-  intl?: {
-    message: object;
   };
 }
